moved app constants in presentation layer and removed some obsolete
