--- app/src/main/java/com/babylon/wallet/android/domain/model/Constants.kt
@@ -1,14 +1,0 @@
-package com.babylon.wallet.android.domain.model
-
-object AppConstants {
-    const val VM_STOP_TIMEOUT_MS = 5000L
-    const val DELAY_300_MS = 300L
-    const val ACCOUNT_NAME_MAX_LENGTH = 20
-}
-
-object MetadataConstants {
-    const val KEY_SYMBOL = "symbol"
-    const val KEY_ICON = "icon_url"
-
-    const val SYMBOL_XRD = "XRD"
-}
