+++ app/src/main/java/com/babylon/wallet/android/utils/Constants.kt
@@ -1,0 +1,7 @@
+package com.babylon.wallet.android.utils
+
+object AppConstants {
+    const val VM_STOP_TIMEOUT_MS = 5000L
+    const val DELAY_300_MS = 300L
+    const val ACCOUNT_NAME_MAX_LENGTH = 20
+}
